smartbench-backend/app/models/types.py
"""Dialect-aware SQLAlchemy column types."""

from __future__ import annotations

from sqlalchemy import JSON
from sqlalchemy.dialects.postgresql import JSONB

JSONType = JSON().with_variant(JSONB, "postgresql")
