smartbench-backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smartbench-backend"
version = "0.1.0"
description = "SmartBench backend platform for scientific data and agent-governed operations"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "SmartBench Engineering" }
]
dependencies = [
  "Flask==3.1.0",
  "Flask-SQLAlchemy==3.1.1",
  "Flask-Migrate==4.0.7",
  "Flask-Login==0.6.3",
  "python-dotenv==1.0.1",
  "pydantic==2.11.1",
  "pydantic-settings==2.8.1",
  "psycopg[binary]==3.2.6",
  "openai==1.68.2",
  "structlog==25.2.0",
  "celery==5.4.0",
  "redis==5.2.1",
  "gunicorn==23.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.5",
  "pytest-cov==6.0.0",
  "ruff==0.11.0",
  "mypy==1.15.0",
  "types-redis==4.6.0.20241004",
  "pre-commit==4.1.0",
]

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-q"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "W"]
ignore = ["E203"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]
